feat(bench): Add basic criterion-based benchmark framework
The benchmarks can be run with `cargo bench` or `cargo criterion`.

I went for criterion because I have some experience with it, and `boa`
also seems to use it for some small engine benchmarks, but it is
possible a different benchmark framework would be more appropriate in
this setting (also for longer running benchmarks).

I only benchmark the execution time, not the parsing of the scripts. The
way I separated the parsing out feels a bit awkward though. In
particular, I am parsing the script for every execution iteration, but I
didn't see another way.

I tried to keep the benchmarked scripts small/fast (in the hundreds of
microseconds range). An exception is `binary-trees.js`, which is meant
to benchmark the GC. I'm unsure how many allocations are needed to
trigger GC. One future possibility is to expose a "perform GC" function
to the benchmarked JS.